Updated Features [Hardware]: - Adding the new Clip Unit stage inbetween the Input Assembler and the Triangle Setup stages. - Replaced the F_to_I23_RoundNearestEven mode on the Float ALU (CNV pipe) with the new F_to_U24_RoundNearestEven mode. - Rewrote most of the Triangle Setup unit to now handle the division by W and viewport transform that the vertex shader used to do. Updated Features [Driver]: - Added the new Clip Unit to the GPU Stats page. Bugfixes [Testbench]: - Fixed a testbench bug where the Triangle Setup testbench wasn't accurately testing the computed Z/W components.

+++ CPPTestbench1/CPPTestbench1/xsi_loader.h
@@ -51,7 +51,7 @@ public:
 
 public:
     // Utility functions
-    int num_ports() { return _xsi_value_buffer.size(); };
+    int num_ports() { return (const int)_xsi_value_buffer.size(); };
     void display_value(int port_number);
     void display_port_values();
     static std::string convert_std_logic_to_str(const char* value_buffer, int width);
